ISPN-10520 Move org.infinispan.Version to org.infinispan.commons.util.Version
* Also remove the main-class entry that pointed to the Version class from all manifests
