+++ commons/src/main/resources/META-INF/infinispan-version.properties
