+++ core/src/main/java/org/infinispan/cache/impl/CacheImpl.java
@@ -32,7 +32,7 @@ import org.infinispan.AdvancedCache;
 import org.infinispan.CacheCollection;
 import org.infinispan.CacheSet;
 import org.infinispan.LockedStream;
-import org.infinispan.Version;
+import org.infinispan.commons.util.Version;
 import org.infinispan.batch.BatchContainer;
 import org.infinispan.commands.CommandsFactory;
 import org.infinispan.commands.VisitableCommand;
