+++ integrationtests/wildfly-modules/src/test/java/org/infinispan/test/integration/as/InfinispanStoreRocksDBIT.java
@@ -6,7 +6,7 @@ import static org.junit.Assert.assertEquals;
 import java.io.File;
 
 import org.infinispan.Cache;
-import org.infinispan.Version;
+import org.infinispan.commons.util.Version;
 import org.infinispan.commons.util.Util;
 import org.infinispan.configuration.cache.ConfigurationBuilder;
 import org.infinispan.configuration.global.GlobalConfigurationBuilder;
