+++ integrationtests/wildfly-modules/src/test/java/org/infinispan/test/integration/as/cdi/GreetingServiceIT.java
@@ -6,7 +6,7 @@ import static org.junit.Assert.assertTrue;
 import javax.cache.annotation.CacheKey;
 import javax.inject.Inject;
 
-import org.infinispan.Version;
+import org.infinispan.commons.util.Version;
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.junit.Arquillian;
 import org.jboss.shrinkwrap.api.Archive;
