+++ integrationtests/wildfly-modules/src/test/java/org/infinispan/test/integration/as/query/InfinispanQueryIT.java
@@ -1,6 +1,6 @@
 package org.infinispan.test.integration.as.query;
 
-import org.infinispan.Version;
+import org.infinispan.commons.util.Version;
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.container.test.api.TargetsContainer;
 import org.jboss.arquillian.junit.Arquillian;
